Move SystemNorm implementations out of header.
